Fixed media delete on media/browser page

+++ src/mvc/public/media/actions/delete.php
@@ -1,7 +1,17 @@
 <?php
 $success = false;
-if ( !empty($ctrl->post['id']) ){
+if (!empty($ctrl->post['id'])) {
 	$medias = new \bbn\Appui\Medias($ctrl->db);
-  $success = $medias->delete($ctrl->post['id']);
+  if (is_array($ctrl->post['id'])) {
+    $success = true;
+    foreach ($ctrl->post['id'] as $id) {
+      if (!$medias->delete($id)) {
+        $success = false;
+      }
+    }
+  }
+  else {
+    $success = $medias->delete($ctrl->post['id']);
+  }
 }
 $ctrl->obj->success = $success;
